Bài tập ss3

+++ src/LocationClass/Matrix.java
@@ -10,18 +10,18 @@ public class Matrix {
         System.out.println("Enter a number of collums");
         int Cols = scan.nextInt();
         createMatrix matrix = new createMatrix();
-        int[][] yourmatrix = matrix.createNewMatrix(Rows,Cols);
+        double[][] yourmatrix = matrix.createNewMatrix(Rows,Cols);
         enterData(scan,yourmatrix,matrix.Rows,matrix.Cols);
-        int[] result = matrix.locateLargest(yourmatrix);
+        double[] result = matrix.locateLargest(yourmatrix);
         System.out.println("Your largest number is " + result[2] +
-                " at row " + result[0] + " and column " + result[1] );
+                " at row " + (int)result[0] + " and column " + (int)result[1] );
     }
-    public static void enterData(Scanner scan,int[][] matrix, int Rows, int Cols){
+    public static void enterData(Scanner scan,double[][] matrix, int Rows, int Cols){
         System.out.println("Enter Matrix Data");
         for(int i = 0;i < Rows; i++){
             System.out.println("Row " + i);
             for (int j = 0; j < Cols; j++){
-                matrix[i][j] = scan.nextInt();
+                matrix[i][j] = scan.nextDouble();
             }
         }
     }
